Last Sync

+++ packer/scripts/ubuntu/base.sh
@@ -8,5 +8,7 @@ message () {
 message "Setting one active console"
 sed -i -e 's,^\(ACTIVE_CONSOLES="/dev/tty\).*,\11",' /etc/default/console-setup
 for f in /etc/init/tty[^1]*.conf; do
-  mv "$f"{,.bak}
+  [[ -e "$f" ]] && mv "$f"{,.bak}
 done
+
+exit 0
